Enemy Strafing, Physics Fix, Independent Projectiles.
An OnCollisionExit function fixed the enemy physics issues stemming from bumping into the player or each other. Created Cross Shot and X Shot and a child class of projectiles to be independent. Finally implemented Enemy Strafing. Might want to set up a slightly faster speed to be used for chasing and strafing.